+++ CHANGELOG.md
@@ -3,7 +3,21 @@
 NOWnews-api change logs
 
 ## [Unreleased]
+### Fixed
+- 修正 google trace engine 造成 CPU 與 Memory 消耗過大問題 @SimonSun
 
+### Added
+- 加入 imagelab 縮圖網址與在 Image mode 加入 sizeFormat 虛擬欄位 @SimonSun
+
+## 1.0.30 - 2017-09-04
+### Fixed
+- 修正 中央社匯入圖片有IP限制 外網看不到的問題 @appleoxxo
+
+## 1.0.29 - 2017-09-01
+### Fixed
+- 修正 ios 內文圖出不來的問題 @webber
+
+## 1.0.28 - 2017-08-31
 ### Changed
 - Admin的api端點 /users 加上 isInitUser 判斷 ＠appleoxxo
 - Adminapi端點 /news 加上 isFeed 判斷 ＠appleoxxo
@@ -15,6 +29,9 @@ NOWnews-api change logs
 ### Added
 - 刪除新聞的時候，也要有 log (因應這個功能，所以在 `newLog` 這個 schema 加入新的欄位 `newsTrashed`) @SimonSun
 - cronjob加上中央社圖片匯入 @appleoxxo
+
+### Changed
+- 優化圖片搜尋增加 index @esbb48
 
 ###Fixed
 - instant 補傳 MainVideo 的欄位 @esbb48